Module aws_sdk_kafkaconnect::error

Errors that can occur when calling the service.

Modules

bad_request_exception

See BadRequestException

conflict_exception

See ConflictException

forbidden_exception

See ForbiddenException

internal_server_error_exception

See InternalServerErrorException

not_found_exception

See NotFoundException

service_unavailable_exception

See ServiceUnavailableException

too_many_requests_exception

See TooManyRequestsException

unauthorized_exception

See UnauthorizedException

Structs

BadRequestException

HTTP Status Code 400: Bad request due to incorrect input. Correct your request and then retry it.

ConflictException

HTTP Status Code 409: Conflict. A resource with this name already exists. Retry your request with another name.

CreateConnectorError

Error type for the CreateConnector operation.

CreateCustomPluginError

Error type for the CreateCustomPlugin operation.

CreateWorkerConfigurationError

Error type for the CreateWorkerConfiguration operation.

DeleteConnectorError

Error type for the DeleteConnector operation.

DescribeConnectorError

Error type for the DescribeConnector operation.

DescribeCustomPluginError

Error type for the DescribeCustomPlugin operation.

DescribeWorkerConfigurationError

Error type for the DescribeWorkerConfiguration operation.

ForbiddenException

HTTP Status Code 403: Access forbidden. Correct your credentials and then retry your request.

InternalServerErrorException

HTTP Status Code 500: Unexpected internal server error. Retrying your request might resolve the issue.

ListConnectorsError

Error type for the ListConnectors operation.

ListCustomPluginsError

Error type for the ListCustomPlugins operation.

ListWorkerConfigurationsError

Error type for the ListWorkerConfigurations operation.

NotFoundException

HTTP Status Code 404: Resource not found due to incorrect input. Correct your request and then retry it.

ServiceUnavailableException

HTTP Status Code 503: Service Unavailable. Retrying your request in some time might resolve the issue.

TooManyRequestsException

HTTP Status Code 429: Limit exceeded. Resource limit reached.

UnauthorizedException

HTTP Status Code 401: Unauthorized request. The provided credentials couldn't be validated.

UpdateConnectorError

Error type for the UpdateConnector operation.

Enums

CreateConnectorErrorKind

Types of errors that can occur for the CreateConnector operation.

CreateCustomPluginErrorKind

Types of errors that can occur for the CreateCustomPlugin operation.

CreateWorkerConfigurationErrorKind

Types of errors that can occur for the CreateWorkerConfiguration operation.

DeleteConnectorErrorKind

Types of errors that can occur for the DeleteConnector operation.

DescribeConnectorErrorKind

Types of errors that can occur for the DescribeConnector operation.

DescribeCustomPluginErrorKind

Types of errors that can occur for the DescribeCustomPlugin operation.

DescribeWorkerConfigurationErrorKind

Types of errors that can occur for the DescribeWorkerConfiguration operation.

ListConnectorsErrorKind

Types of errors that can occur for the ListConnectors operation.

ListCustomPluginsErrorKind

Types of errors that can occur for the ListCustomPlugins operation.

ListWorkerConfigurationsErrorKind

Types of errors that can occur for the ListWorkerConfigurations operation.

UpdateConnectorErrorKind

Types of errors that can occur for the UpdateConnector operation.