Module aws_sdk_kafkaconnect::model

Data structures used by operation inputs/outputs.

Modules

apache_kafka_cluster

See ApacheKafkaCluster

apache_kafka_cluster_description

See ApacheKafkaClusterDescription

auto_scaling

See AutoScaling

auto_scaling_description

See AutoScalingDescription

auto_scaling_update

See AutoScalingUpdate

capacity

See Capacity

capacity_description

See CapacityDescription

capacity_update

See CapacityUpdate

cloud_watch_logs_log_delivery

See CloudWatchLogsLogDelivery

cloud_watch_logs_log_delivery_description

See CloudWatchLogsLogDeliveryDescription

connector_summary

See ConnectorSummary

custom_plugin

See CustomPlugin

custom_plugin_description

See CustomPluginDescription

custom_plugin_file_description

See CustomPluginFileDescription

custom_plugin_location

See CustomPluginLocation

custom_plugin_location_description

See CustomPluginLocationDescription

custom_plugin_revision_summary

See CustomPluginRevisionSummary

custom_plugin_summary

See CustomPluginSummary

firehose_log_delivery

See FirehoseLogDelivery

firehose_log_delivery_description

See FirehoseLogDeliveryDescription

kafka_cluster

See KafkaCluster

kafka_cluster_client_authentication

See KafkaClusterClientAuthentication

kafka_cluster_client_authentication_description

See KafkaClusterClientAuthenticationDescription

kafka_cluster_description

See KafkaClusterDescription

kafka_cluster_encryption_in_transit

See KafkaClusterEncryptionInTransit

kafka_cluster_encryption_in_transit_description

See KafkaClusterEncryptionInTransitDescription

log_delivery

See LogDelivery

log_delivery_description

See LogDeliveryDescription

plugin

See Plugin

plugin_description

See PluginDescription

provisioned_capacity

See ProvisionedCapacity

provisioned_capacity_description

See ProvisionedCapacityDescription

provisioned_capacity_update

See ProvisionedCapacityUpdate

s3_location

See S3Location

s3_location_description

See S3LocationDescription

s3_log_delivery

See S3LogDelivery

s3_log_delivery_description

See S3LogDeliveryDescription

scale_in_policy

See ScaleInPolicy

scale_in_policy_description

See ScaleInPolicyDescription

scale_in_policy_update

See ScaleInPolicyUpdate

scale_out_policy

See ScaleOutPolicy

scale_out_policy_description

See ScaleOutPolicyDescription

scale_out_policy_update

See ScaleOutPolicyUpdate

vpc

See Vpc

vpc_description

See VpcDescription

worker_configuration

See WorkerConfiguration

worker_configuration_description

See WorkerConfigurationDescription

worker_configuration_revision_description

See WorkerConfigurationRevisionDescription

worker_configuration_revision_summary

See WorkerConfigurationRevisionSummary

worker_configuration_summary

See WorkerConfigurationSummary

worker_log_delivery

See WorkerLogDelivery

worker_log_delivery_description

See WorkerLogDeliveryDescription

Structs

ApacheKafkaCluster

The details of the Apache Kafka cluster to which the connector is connected.

ApacheKafkaClusterDescription

The description of the Apache Kafka cluster to which the connector is connected.

AutoScaling

Specifies how the connector scales.

AutoScalingDescription

Information about the auto scaling parameters for the connector.

AutoScalingUpdate

The updates to the auto scaling parameters for the connector.

Capacity

Information about the capacity of the connector, whether it is auto scaled or provisioned.

CapacityDescription

A description of the connector's capacity.

CapacityUpdate

The target capacity for the connector. The capacity can be auto scaled or provisioned.

CloudWatchLogsLogDelivery

The settings for delivering connector logs to Amazon CloudWatch Logs.

CloudWatchLogsLogDeliveryDescription

A description of the log delivery settings.

ConnectorSummary

Summary of a connector.

CustomPlugin

A plugin is an AWS resource that contains the code that defines a connector's logic.

CustomPluginDescription

Details about a custom plugin.

CustomPluginFileDescription

Details about a custom plugin file.

CustomPluginLocation

Information about the location of a custom plugin.

CustomPluginLocationDescription

Information about the location of a custom plugin.

CustomPluginRevisionSummary

Details about the revision of a custom plugin.

CustomPluginSummary

A summary of the custom plugin.

FirehoseLogDelivery

The settings for delivering logs to Amazon Kinesis Data Firehose.

FirehoseLogDeliveryDescription

A description of the settings for delivering logs to Amazon Kinesis Data Firehose.

KafkaCluster

The details of the Apache Kafka cluster to which the connector is connected.

KafkaClusterClientAuthentication

The client authentication information used in order to authenticate with the Apache Kafka cluster.

KafkaClusterClientAuthenticationDescription

The client authentication information used in order to authenticate with the Apache Kafka cluster.

KafkaClusterDescription

Details of how to connect to the Apache Kafka cluster.

KafkaClusterEncryptionInTransit

Details of encryption in transit to the Apache Kafka cluster.

KafkaClusterEncryptionInTransitDescription

The description of the encryption in transit to the Apache Kafka cluster.

LogDelivery

Details about log delivery.

LogDeliveryDescription

The description of the log delivery settings.

Plugin

A plugin is an AWS resource that contains the code that defines your connector logic.

PluginDescription

The description of the plugin.

ProvisionedCapacity

Details about a connector's provisioned capacity.

ProvisionedCapacityDescription

The description of a connector's provisioned capacity.

ProvisionedCapacityUpdate

An update to a connector's fixed capacity.

S3Location

The location of an object in Amazon S3.

S3LocationDescription

The description of the location of an object in Amazon S3.

S3LogDelivery

Details about delivering logs to Amazon S3.

S3LogDeliveryDescription

The description of the details about delivering logs to Amazon S3.

ScaleInPolicy

The scale-in policy for the connector.

ScaleInPolicyDescription

The description of the scale-in policy for the connector.

ScaleInPolicyUpdate

An update to the connector's scale-in policy.

ScaleOutPolicy

The scale-out policy for the connector.

ScaleOutPolicyDescription

The description of the scale-out policy for the connector.

ScaleOutPolicyUpdate

An update to the connector's scale-out policy.

Vpc

Information about the VPC in which the connector resides.

VpcDescription

The description of the VPC in which the connector resides.

WorkerConfiguration

The configuration of the workers, which are the processes that run the connector logic.

WorkerConfigurationDescription

The description of the worker configuration.

WorkerConfigurationRevisionDescription

The description of the worker configuration revision.

WorkerConfigurationRevisionSummary

The summary of a worker configuration revision.

WorkerConfigurationSummary

The summary of a worker configuration.

WorkerLogDelivery

Workers can send worker logs to different destination types. This configuration specifies the details of these destinations.

WorkerLogDeliveryDescription

Workers can send worker logs to different destination types. This configuration specifies the details of these destinations.

Enums

ConnectorState

CustomPluginContentType

CustomPluginState

KafkaClusterClientAuthenticationType

KafkaClusterEncryptionInTransitType