Struct aws_sdk_kafkaconnect::output::DescribeConnectorOutput

#[non_exhaustive]
pub struct DescribeConnectorOutput {

    pub capacity: Option<CapacityDescription>,
    pub connector_arn: Option<String>,
    pub connector_configuration: Option<HashMap<String, String>>,
    pub connector_description: Option<String>,
    pub connector_name: Option<String>,
    pub connector_state: Option<ConnectorState>,
    pub creation_time: Option<Instant>,
    pub current_version: Option<String>,
    pub kafka_cluster: Option<KafkaClusterDescription>,
    pub kafka_cluster_client_authentication: Option<KafkaClusterClientAuthenticationDescription>,
    pub kafka_cluster_encryption_in_transit: Option<KafkaClusterEncryptionInTransitDescription>,
    pub kafka_connect_version: Option<String>,
    pub log_delivery: Option<LogDeliveryDescription>,
    pub plugins: Option<Vec<PluginDescription>>,
    pub service_execution_role_arn: Option<String>,
    pub worker_configuration: Option<WorkerConfigurationDescription>,

}

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

capacity: Option<CapacityDescription>

Information about the capacity of the connector, whether it is auto scaled or provisioned.

connector_arn: Option<String>

The Amazon Resource Name (ARN) of the connector.

connector_configuration: Option<HashMap<String, String>>

A map of keys to values that represent the configuration for the connector.

connector_description: Option<String>

A summary description of the connector.

connector_name: Option<String>

The name of the connector.

connector_state: Option<ConnectorState>

The state of the connector.

creation_time: Option<Instant>

The time the connector was created.

current_version: Option<String>

The current version of the connector.

kafka_cluster: Option<KafkaClusterDescription>

The Apache Kafka cluster that the connector is connected to.

kafka_cluster_client_authentication: Option<KafkaClusterClientAuthenticationDescription>

The type of client authentication used to connect to the Apache Kafka cluster. The value is NONE when no client authentication is used.

kafka_cluster_encryption_in_transit: Option<KafkaClusterEncryptionInTransitDescription>

Details of encryption in transit to the Apache Kafka cluster.

kafka_connect_version: Option<String>

The version of Kafka Connect. It has to be compatible with both the Apache Kafka cluster's version and the plugins.

log_delivery: Option<LogDeliveryDescription>

Details about delivering logs to Amazon CloudWatch Logs.

plugins: Option<Vec<PluginDescription>>

Specifies which plugins were used for this connector.

service_execution_role_arn: Option<String>

The Amazon Resource Name (ARN) of the IAM role used by the connector to access Amazon Web Services resources.

worker_configuration: Option<WorkerConfigurationDescription>

Specifies which worker configuration was used for the connector.

Implementations

impl DescribeConnectorOutput

pub fn capacity(&self) -> Option<&CapacityDescription>

Information about the capacity of the connector, whether it is auto scaled or provisioned.

pub fn connector_arn(&self) -> Option<&str>

The Amazon Resource Name (ARN) of the connector.

pub fn connector_configuration(&self) -> Option<&HashMap<String, String>>

A map of keys to values that represent the configuration for the connector.

pub fn connector_description(&self) -> Option<&str>

A summary description of the connector.

pub fn connector_name(&self) -> Option<&str>

The name of the connector.

pub fn connector_state(&self) -> Option<&ConnectorState>

The state of the connector.

pub fn creation_time(&self) -> Option<&Instant>

The time the connector was created.

pub fn current_version(&self) -> Option<&str>

The current version of the connector.

pub fn kafka_cluster(&self) -> Option<&KafkaClusterDescription>

The Apache Kafka cluster that the connector is connected to.

pub fn kafka_cluster_client_authentication(
&self
) -> Option<&KafkaClusterClientAuthenticationDescription>

The type of client authentication used to connect to the Apache Kafka cluster. The value is NONE when no client authentication is used.

pub fn kafka_cluster_encryption_in_transit(
&self
) -> Option<&KafkaClusterEncryptionInTransitDescription>

Details of encryption in transit to the Apache Kafka cluster.

pub fn kafka_connect_version(&self) -> Option<&str>

The version of Kafka Connect. It has to be compatible with both the Apache Kafka cluster's version and the plugins.

pub fn log_delivery(&self) -> Option<&LogDeliveryDescription>

Details about delivering logs to Amazon CloudWatch Logs.

pub fn plugins(&self) -> Option<&[PluginDescription]>

Specifies which plugins were used for this connector.

pub fn service_execution_role_arn(&self) -> Option<&str>

The Amazon Resource Name (ARN) of the IAM role used by the connector to access Amazon Web Services resources.

pub fn worker_configuration(&self) -> Option<&WorkerConfigurationDescription>

Specifies which worker configuration was used for the connector.

impl DescribeConnectorOutput

pub fn builder() -> Builder

Creates a new builder-style object to manufacture DescribeConnectorOutput